decoupling db from block
