+++ oagdedupe/block/pairs.py
@@ -15,73 +15,22 @@ from sqlalchemy import create_engine
 
 from oagdedupe import utils as du
 from oagdedupe._typing import ENGINE, StatsDict
-from oagdedupe.block.mixin import ConjunctionMixin
+from oagdedupe.base import BaseCompute
 from oagdedupe.containers import Container
 from oagdedupe.settings import Settings
 
 
 @dataclass
-class Pairs(ConjunctionMixin):
+class Pairs:
     """
     Computes pairs for conjunctions and appends to comparisons or
     full_comparisons table.
     """
 
     settings: Settings = Provide[Container.settings]
+    compute: BaseCompute = Provide[Container.blocking]
 
-    def _get_n_pairs(self, table: str) -> pd.DataFrame:
-        newtable = self.comptab_map[table]
-        return self.query(
-            f"""
-            SELECT count(*) FROM {self.settings.db.db_schema}.{newtable}
-        """
-        )["count"].values[0]
-
-    @du.recordlinkage
-    def _save_comparison_pairs(
-        self, names: Tuple[str], table: str, rl: str = ""
-    ) -> None:
-        """
-        Given forward index, construct inverted index.
-        Then for each row in inverted index, get all "nC2" distinct
-        combinations of size 2 from the array.
-
-        Concatenates and returns all distinct pairs.
-
-        Parameters
-        ----------
-        names : List[str]
-            list of block schemes
-        table : str
-            table name of forward index
-
-        Returns
-        ----------
-        pd.DataFrame
-        """
-        newtable = self.comptab_map[table]
-        engine = create_engine(self.settings.db.path_database)
-        engine.execute(
-            f"""
-            INSERT INTO {self.settings.db.db_schema}.{newtable}
-            (
-                WITH
-                    inverted_index AS (
-                        {self._inv_idx_query(names, table)}
-                    ),
-                    inverted_index_link AS (
-                        {self._inv_idx_query(names, table+rl, col="_index_r")}
-                    )
-                {self._pairs_query(names)}
-            )
-            ON CONFLICT DO NOTHING
-            """
-        )
-        engine.dispose()
-
-    def add_new_comparisons(
-        self, stats: StatsDict, table: str, engine: ENGINE
-    ) -> int:
+    def add_new_comparisons(self, stats: StatsDict, table: str) -> int:
         """
         Computes pairs for conjunction and appends to comparisons or
         full_comparisons table.
@@ -105,7 +54,6 @@ class Pairs(ConjunctionMixin):
         int
             total number of pairs gathered so far
         """
-
-        self._save_comparison_pairs(names=stats.scheme, table=table)
-        n = self._get_n_pairs(table=table)
+        self.compute._save_comparison_pairs(names=stats.scheme, table=table)
+        n = self.compute._get_n_pairs(table=table)
         return n
