+++ oagdedupe/containers.py
@@ -1,6 +1,6 @@
 from dependency_injector import containers, providers
 
-from oagdedupe.base import BaseCompute
+from oagdedupe.base import BaseCompute, BaseComputeBlocking
 from oagdedupe.settings import Settings
 
 
@@ -9,6 +9,7 @@ class Container(containers.DeclarativeContainer):
     wiring_config = containers.WiringConfiguration(
         packages=[
             "oagdedupe.db",
+            "oagdedupe.db.postgres",
             "oagdedupe.block",
             "oagdedupe.distance",
             "oagdedupe.cluster",
@@ -18,3 +19,5 @@ class Container(containers.DeclarativeContainer):
     settings = providers.Factory(Settings)
 
     compute = providers.Factory(BaseCompute)
+
+    blocking = providers.Factory(BaseComputeBlocking)
