+++ oagdedupe/db/postgres/compute.py
@@ -1,6 +1,7 @@
 import itertools
 import logging
 from dataclasses import dataclass
+from functools import cached_property
 from typing import List
 
 from dependency_injector.wiring import Provide
@@ -8,8 +9,9 @@ from sqlalchemy import delete, func, select
 
 from oagdedupe import utils as du
 from oagdedupe._typing import SESSION, TABLE
-from oagdedupe.base import BaseCompute
+from oagdedupe.base import BaseCompute, BaseComputeBlocking
 from oagdedupe.containers import Container
+from oagdedupe.db.postgres.blocking import PostgresBlocking
 from oagdedupe.db.postgres.initialize import Initialize
 from oagdedupe.db.postgres.orm import DatabaseORM
 from oagdedupe.db.postgres.tables import Tables
@@ -23,4 +25,6 @@ class PostgresCompute(BaseCompute, Initialize, DatabaseORM, Tables):
 
     settings: Settings = Provide[Container.settings]
 
-    pass
+    @cached_property
+    def blocking(self) -> BaseComputeBlocking:
+        return PostgresBlocking()
